requirements.txt
Django==1.8.5
MySQL-python==1.2.5
Pillow==3.0.0
argparse==1.2.1
django-crispy-forms==1.5.2
docopt==0.6.2
psycopg2==2.6.1
python-util==1.2.1
wsgiref==0.1.2
